Template update (#16)
* first try at updating the template (in-progress)

* fixed most major issues some concerns will be in PR

* small fix for production env

* added a propper production build pipeline

* npm install for service will now only run on build

* fixed production build so that npm install doesn't run on every build, fixed relative file paths for FS

* fixed merge conflicts

+++ boot.sh
@@ -1,12 +1,18 @@
 #! /bin/sh
-#remove app folder if exists
-rm -rf ./app;
-#copy app folder
-cp -r /app ./;
-#install app dependencies
-npm install ./app;
-#remove package to avoid babel and imports breaking
-#TODO: if there are nested package.json files stuff will break but I think this is true for older versions too
-rm ./app/package.json;
-#run daemon and watch mounted app folder
-npm run daemon;
+if [ "$NODE_ENV" == "development" ]
+then
+  #remove app folder if exists
+  rm -rf ./app;
+  #copy app folder
+  cp -r /app ./;
+  #install app dependencies
+  npm install ./app;
+  #remove package to avoid babel and imports breaking
+  #TODO: if there are nested package.json files stuff will break but I think this is true for older versions too
+  rm ./app/package.json;
+  #run daemon and watch mounted app folder
+  npm run daemon;
+else
+  #run production app
+  npm run node-prod;
+fi
